Match several particle functions (#2576)
* hsd_80393D2C: 94.34% -> 100% (match=true)

Access hsd_804CF7E8 via a local struct pointer alias instead of the
global directly. This makes MWCC load &hsd_804CF7E8 into a register
(lbzu form, offset-0 accesses) rather than using the .bss section
anchor + offset 260, which had cost an extra addi instruction.


* particle: match hsd_803966A0, hsd_80396C78, fn_80397374 (3 fns -> 100%)

Two levers:
1. ps_remove_node inline: fold head_ptr assignment into the NULL check
   (if (*(head_ptr = &sp->xD0) == NULL)) so MWCC emits the addi (head_ptr
   compute) before the lwz, matching the original instruction schedule.
   This alone matched hsd_80396C78 and improved hsd_80395D88/hsd_803962A8/
   hsd_80396A20.
2. Access hsd_804CF810 through a local struct-pointer alias (sp) instead of
   (&hsd_804CF810)-> so MWCC references the named symbol (offset-0 base)
   rather than the .bss section anchor + offset. Matched hsd_803966A0 and
   fn_80397374 once combined with lever 1.

Removed stale @TODO match-percent annotations on the matched functions.


* Match hsd_8039EFAC and hsd_8039F6CC; correct hsd_8039F05C signature

The hsd_8039F05C prototype declared its first parameter as s8, forcing a
narrowing (extsb) at callers that pass a full-word linkNo which the target
binary does not have. Correcting it to s32 -- consistent with the int linkNo
convention used throughout psstructs.h -- removes the spurious narrowing in
hsd_8039EFAC, hsd_8039F6CC, and hsd_8039F05C itself.

With the signature corrected, hsd_8039EFAC and hsd_8039F6CC match by
reproducing the target control flow: the jobj store and inlined ref_INC are
guarded by explicit gen/jobj null-checks, which produce the redundant (dead)
branches present in the target binary.

efLib_CreateGenerator's linkNo local is widened u8 -> int so its register
allocation continues to match under the corrected s32 parameter.

hsd_8039F05C improves (80.55 -> 80.79) but remains unmatched: a .bss section
anchor hoist drives a register-allocation cascade.

Matched: hsd_8039EFAC, hsd_8039F6CC


---------

+++ src/melee/ef/eflib.c
@@ -682,7 +682,7 @@ void efLib_particles_proc_aux(HSD_GObj* gobj)
 HSD_Generator* efLib_CreateGenerator(s32 gfx_id, Vec3* pos)
 {
     HSD_Generator* generator;
-    u8 linkNo;
+    int linkNo;
 
     linkNo = 0;
     switch (gfx_id) {
